[codex] Sync upstream and TheTom TurboQuant KV cache

+++ tools/perplexity/perplexity.cpp
@@ -524,7 +524,12 @@ static results_perplexity perplexity(llama_context * ctx, const common_params & 
         logits_stream.write((const char *)&n_chunk, sizeof(n_chunk));
         logits_stream.write((const char *)tokens.data(), n_chunk*n_ctx*sizeof(tokens[0]));
         const int nv = 2*((n_vocab + 1)/2) + 4;
-        log_probs.resize(n_ctx * nv);
+        // size_t cast: int * int overflows on Qwen-class large-vocab models at
+        // n_ctx >= 16K (e.g. n_ctx=16384 * nv=151940 = 2.49B > INT32_MAX=2.15B);
+        // overflow wraps negative, sign-extends to a giant size_t when passed to
+        // resize(), trips vector::max_size and throws std::length_error. Matches
+        // the existing size_t cast on line 514 above for the same reason.
+        log_probs.resize(size_t(n_ctx) * nv);
     }
 
     // We get the logits for all the tokens in the context window (params.n_ctx)
